api/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
google-genai>=1.0.0
google-cloud-vision>=3.8.0
pydantic>=2.10.0
Pillow>=11.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
